Migrate zksend sdk methods to sui sdk
