+++ sdk/docs/pages/typescript/index.mdx
@@ -107,5 +107,4 @@ what you need to keep your code light and compact.
   signatures.
 - [`@mysten/sui/utils`](/typescript/utils) - Utilities for formatting and parsing various Sui types.
 - [`@mysten/sui/faucet`](/typescript/faucet) - Methods for requesting SUI from a faucet.
-- [`@mysten/sui/zklogin`](/typescript/zklogin) - Utilities for working with zkLogin. (Note
-  @mysten/zklogin package also contains zklogin related functionality)
+- [`@mysten/sui/zklogin`](/typescript/zklogin) - Utilities for working with zkLogin.
