+++ sdk/docs/pages/typescript/zklogin.mdx
@@ -5,10 +5,6 @@ import { Callout } from 'nextra/components';
 Utilities for working with zkLogin. Currently contains functionality to create and parse zkLogin
 signatures and compute zkLogin addresses.
 
-<Callout type="info">
-	Note: `@mysten/zklogin` package contains more utilities to help with computing address etc.
-</Callout>
-
 To parse a serialized zkLogin signature
 
 ```typescript
@@ -30,8 +26,63 @@ To compute the address for a given address seed and iss you can use `computeZkLo
 ```typescript
 import { computeZkLoginAddressFromSeed } from '@mysten/sui/zklogin';
 
-const address = await computeZkLoginAddressFromSeed(0n, 'https://accounts.google.com');
+const address = computeZkLoginAddressFromSeed(0n, 'https://accounts.google.com');
+```
+
+To compute an address from jwt:
+
+```typescript
+import { jwtToAddress } from '@mysten/sui/zklogin';
+
+const address = jwtToAddress(jwtAsString, salt);
+```
+
+To compute an address from a parsed jwt:
+
+```typescript
+import { computeZkLoginAddress } from '@mysten/sui/zklogin';
+
+const address = computeZkLoginAddress({
+	claimName,
+	claimValue,
+	iss,
+	aud,
+	userSalt: BigInt(salt),
+});
 ```
 
 To use zkLogin inside a multisig, see the [Multisig Guide](../typescript/cryptography/multisig.mdx)
 for more details.
+
+## Legacy addresses
+
+When zklogin was first introduced, there was an inconsistency in how the address seed was computed.
+For backwards compatibility reasons there are 2 valid addresses for a given set of inputs. Methods
+that produce zklogin addresses all accept a `legacyAddress` boolean flag, either as their last
+parameter, or in their options argument.
+
+```typescript
+import {
+	computeZkLoginAddress,
+	computeZkLoginAddressFromSeed,
+	jwtToAddress,
+	toZkLoginPublicIdentifier,
+	genAddressSeed(userSalt, claimName, claimValue, aud)
+} from '@mysten/sui/zklogin';
+
+const address = jwtToAddress(jwtAsString, salt, true);
+const address = computeZkLoginAddressFromSeed(0n, 'https://accounts.google.com', true);
+const address = computeZkLoginAddress({
+	claimName,
+	claimValue,
+	iss,
+	aud,
+	userSalt: BigInt(salt),
+	legacyAddress: true,
+});
+const address = toZkLoginPublicIdentifier(
+	genAddressSeed(userSalt, claimName, claimValue, aud),
+	iss,
+	{ legacyAddress: true }
+).toSuiAddress();
+```
