+++ sdk/enoki/src/EnokiKeypair.ts
@@ -5,7 +5,7 @@ import type { SignatureWithBytes } from '@mysten/sui/cryptography';
 import { Keypair, PublicKey, SIGNATURE_SCHEME_TO_FLAG } from '@mysten/sui/cryptography';
 import type { Ed25519Keypair, Ed25519PublicKey } from '@mysten/sui/keypairs/ed25519';
 import type { ZkLoginSignatureInputs } from '@mysten/sui/zklogin';
-import { getZkLoginSignature } from '@mysten/zklogin';
+import { getZkLoginSignature } from '@mysten/sui/zklogin';
 
 export class EnokiPublicKey extends PublicKey {
 	#address: string;
