+++ sdk/typescript/src/zklogin/index.ts
@@ -2,7 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export { getZkLoginSignature, parseZkLoginSignature } from './signature.js';
-export { toBigEndianBytes, toPaddedBigEndianBytes } from './utils.js';
-export { computeZkLoginAddressFromSeed } from './address.js';
+export {
+	toBigEndianBytes,
+	toPaddedBigEndianBytes,
+	hashASCIIStrToField,
+	genAddressSeed,
+	getExtendedEphemeralPublicKey,
+} from './utils.js';
+export { computeZkLoginAddressFromSeed, computeZkLoginAddress, jwtToAddress } from './address.js';
+export type { ComputeZkLoginAddressOptions } from './address.js';
 export { toZkLoginPublicIdentifier, ZkLoginPublicIdentifier } from './publickey.js';
 export type { ZkLoginSignatureInputs } from './bcs.js';
+export { poseidonHash } from './poseidon.js';
+export { generateNonce, generateRandomness } from './nonce.js';
