+++ sdk/zklogin/README.md
@@ -1,4 +1,4 @@
-# `@mysten/zklogin`
+# `@mysten/zklogin` (Deprecated, use @mysten/sui/zklogin instead)
 
 > **⚠️ This package is experimental and will change rapidly as it is being developed. Do not
 > consider these APIs to be stable. If you have any feedback,
